Move encoded script logic together
