+++ app/web/src/helpers/hooks/useIdeContext.ts
@@ -1,0 +1,6 @@
+import { IdeContext } from 'src/pages/DevIdePage/DevIdePage'
+import { useContext } from 'react'
+
+export function useIdeContext() {
+  return useContext(IdeContext)
+}
